Write unit tests for Android version code
