+++ .github/actions/bumpVersion/bumpVersion.js
@@ -4,10 +4,7 @@ const fs = require('fs');
 const core = require('@actions/core');
 const github = require('@actions/github');
 const semverClean = require('semver/functions/clean');
-const getMajorVersion = require('semver/functions/major');
-const getMinorVersion = require('semver/functions/minor');
-const getPatchVersion = require('semver/functions/patch');
-const getBuildVersion = require('semver/functions/prerelease');
+const generateAndroidVersionCode = require('../../libs/generateAndroidVersionCode');
 
 // Filepath constants
 const BUILD_GRADLE_PATH = './android/app/build.gradle';
@@ -16,39 +13,6 @@ const PLIST_PATH_TEST = './ios/ExpensifyCashTests/Info.plist';
 
 // Promisified version of fs.readFile
 const readFileAsync = promisify(fs.readFile);
-
-/**
- * Pad a number to be three digits (with leading zeros if necessary).
- *
- * @param {Number} number - Must be an integer.
- * @returns {String} - A string representation of the number w/ length 3.
- */
-function padToThreeDigits(number) {
-    if (number >= 100) {
-        return number.toString();
-    }
-    if (number >= 10) {
-        return `0${number.toString()}`;
-    }
-    return `00${number.toString()}`;
-}
-
-/**
- * Generate the 12-digit versionCode for android.
- * This version code allocates three digits each for MAJOR, MINOR, PATCH, and BUILD versions.
- * As a result, our max version is 999.999.999-999.
- *
- * @param {String} npmVersion
- * @returns {String}
- */
-function generateAndroidVersionCode(npmVersion) {
-    return ''.concat(
-        padToThreeDigits(getMajorVersion(npmVersion)),
-        padToThreeDigits(getMinorVersion(npmVersion)),
-        padToThreeDigits(getPatchVersion(npmVersion)),
-        padToThreeDigits(getBuildVersion(npmVersion)),
-    );
-}
 
 /**
  * Update the Android app version.
